WayPropertySetSourceFactory extracted as a separate class
